pyproject.toml
[tool.poetry]
name = "koc_tournament_api"
version = "0.1.0"
description = ""
authors = ["9AI"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
pydantic = {extras = ["email"], version = "^2.6.3"}
uvicorn = "^0.27.1"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
pymongo = "^4.6.2"
loguru = "^0.7.2"
motor = "^3.3.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
orjson = "^3.9.15"
bcrypt = "^4.1.2"
cloudinary = "^1.40.0"


[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.23.5"
pytest = "^8.0.2"
httpx = "^0.23.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.2.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
autoflake = "^2.3.0"
pre-commit = "^3.6.2"

[tool.black]
line-length = 90
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
